+++ src/components/TextField/TextField.tsx
@@ -7,7 +7,7 @@ import { st, classes } from './TextField.st.css';
 import { ErrorProps } from '../ErrorMessageWrapper';
 import { ReactComponent as ErrorIcon } from '../../assets/icons/Error.svg';
 import { ReactComponent as SuccessIcon } from '../../assets/icons/CheckSuccess.svg';
-import { ReactComponent as ClearIcon } from './ClearIcon.svg';
+import { ClearIcon } from './ClearIcon';
 import { IconButton, Skins } from '../IconButton';
 import { Tooltip } from '../Tooltip';
 import { TooltipSkin } from '../Tooltip/TooltipEnums';
@@ -25,6 +25,8 @@ import { TPAComponentProps } from '../../types';
 export interface TPATextFieldProps extends TPAComponentProps {
   /** the error message to display */
   errorMessage?: string;
+  /** error tooltip max width */
+  errorTooltipMaxWidth?: number;
   /** possible values: 'line', 'box' */
   theme?: TextFieldTheme;
   /** apply success state */
@@ -87,6 +89,7 @@ export class TextField extends React.Component<TextFieldProps> {
       onClear,
       value,
       disabled,
+      errorTooltipMaxWidth,
     } = this.props;
 
     const shouldShowCustomSuffix = !!suffix;
@@ -127,8 +130,16 @@ export class TextField extends React.Component<TextFieldProps> {
           errorMessage={errorMessage}
           success={success}
           successIcon={successIcon}
+          errorTooltipMaxWidth={errorTooltipMaxWidth}
         />
-        {suffix && <div className={classes.customSuffix} data-hook={DATA_HOOKS.CUSTOM_SUFFIX}>{suffix}</div>}
+        {suffix && (
+          <div
+            className={classes.customSuffix}
+            data-hook={DATA_HOOKS.CUSTOM_SUFFIX}
+          >
+            {suffix}
+          </div>
+        )}
       </div>
     ) : null;
   };
@@ -181,23 +192,35 @@ export class TextField extends React.Component<TextFieldProps> {
   }
 }
 
-const ErrorSuffix = ({ errorMessage }) => (
+const ErrorSuffix = ({ errorMessage, tooltipMaxWidth }) => (
   <Tooltip
     appendTo="scrollParent"
     placement="top-end"
     skin={TooltipSkin.Error}
     content={errorMessage}
     moveBy={{ x: 5, y: 0 }}
+    maxWidth={tooltipMaxWidth}
   >
     <ErrorIcon />
   </Tooltip>
 );
 
-const StatusIcon = ({ error, errorMessage, success, successIcon }) => {
+const StatusIcon = ({
+  error,
+  errorMessage,
+  success,
+  successIcon,
+  errorTooltipMaxWidth,
+}) => {
   let statusIcon = null;
 
   if (errorMessage && error) {
-    statusIcon = <ErrorSuffix errorMessage={errorMessage} />;
+    statusIcon = (
+      <ErrorSuffix
+        errorMessage={errorMessage}
+        tooltipMaxWidth={errorTooltipMaxWidth}
+      />
+    );
   } else if (successIcon && success) {
     statusIcon = <SuccessIcon data-hook={DATA_HOOKS.SUCCESS_ICON} />;
   }